Move mirror.resource service to ldp.remote
