+++ src/middleware/packages/ldp/service.js
@@ -1,7 +1,8 @@
 const LdpContainerService = require('./services/container');
-const LdpResourceService = require('./services/resource');
 const LdpCacheService = require('./services/cache');
 const LdpRegistryService = require('./services/registry');
+const LdpRemoteService = require('./services/remote');
+const LdpResourceService = require('./services/resource');
 
 module.exports = {
   name: 'ldp',
