+++ src/middleware/packages/ldp/services/remote/actions/getNetwork.js
